+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+### 4.8.12
+
+- feat: timebar supports individual container DOM;
+
+### 4.8.11-beta.0
+
+- fix: position animations;
+
 ### 4.8.10
 
 - fix: update layout from animate false to true;